refactor: move user_profile app into users (#6)
