+++ mypong/mypong/settings.py
@@ -41,8 +41,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
-    'users',
-    'user_profile',
+    'users'
 ]
 
 MIDDLEWARE = [
