--- mypong/user_profile/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from .views import UserProfileAPI
-
-app_name = 'user_profile'
-
-urlpatterns = [
-    path('profile/', UserProfileAPI.as_view()),
-]
